Strip all newlines and newline dashes in block
This represents a significant shift in approach towards dealing with
dashes. They are difficult because some species names have real dashes
in them, but preserving dashes causes problems with parsing and results
in a number of species names that have inappropriate dashes in addition
to the appropriate ones so doesn't really solve the problem anyway.

We will handle putting the dashes back in the species names using fuzzy
matching at the end.