Allow varying margins to be set for OCR
Different page layouts from different journals/years require the ability
to change the cropping region for OCR to remove headers and footers
prior to processing.

This change uses this new feature to properly OCR the 2006 data, which
needs to be OCR'd because the column alignments are broken starting
~2/3's of the way through the document resulting in pieces of two
different columns being intermingled.